requirements.txt
fastapi
uvicorn
pymupdf
sentence-transformers
numpy
scikit-learn
openai
httpx 
python-dotenv
streamlit
fitz
